Add ghost effect to fragment shader with blue tint and transparency
- Introduced a new uniform variable `isGhost` to control ghost rendering.
- Implemented ghost placement preview with a blue tint and semi-transparency when `isGhost` is true.
- Maintained existing selection highlight functionality for selected objects.
-Also tweaked the combat system to support "factions" by adding a tag, now ships automatically attack enemies and also you can manualy click on them if they are in range to attack

+++ GuiManager.hpp
@@ -123,7 +123,7 @@ namespace gps {
         bool ColoredButton(const char* label, ImVec2 size, ImVec4 color, ImVec4 hoverColor);
 
         int m_tileSize;
-        float m_tileModelScale;
+        int m_tileModelScale;
 
         int heightTile;
     };
